clippy: needless_return, unnecessary_cast, redundant statics/locals
Zero-judgment mechanical cleanups:
- needless_return (4): silk_sigm_q15 branch returns become tail
  expressions (same branches, expression form).
- unnecessary_cast (3): drop no-op casts — the outer frac_mul16(..) as i32
  in bitexact_cos (frac_mul16 already returns i32), len as i32 (len is
  c_int), dec.storage as u32 (storage is u32).
- redundant_static_lifetimes (4): &'static -> & in the SILK table
  pointer statics ('static is the default there).
- redundant_locals (1): drop the no-op `let len = len;`.

Held back (not auto-applied): needless_late_init (sits in the C-style
grouped declarations of quant_band), identity_op (a deliberate `+ 0`
paralleling `+ 2` on the next line), explicit_counter_loop /
needless_range_loop (loop-idiom family — same open question), and the
control-flow lints collapsible_if / if_same_then_else (brush against the
"don't restructure C control flow" rule). needless_option_as_deref left
too: its fix turns a reborrow into a move — real semantic change, marginal
gain.

Vectors bit-exact (97.0750 / 97.7667); 36 tests pass on both feature sets.

+++ src/bands.rs
@@ -261,8 +261,7 @@ const QTHETA_OFFSET_TWOPHASE: c_int = 16;
 fn bitexact_cos(x: i16) -> i16 {
     let tmp = (4096 + (x as i32) * (x as i32)) >> 13;
     let x2 = tmp as i16;
-    let x2 =
-        (32767 - x2 as i32 + frac_mul16(x2, -7651 + frac_mul16(x2, 8277 + frac_mul16(-626, x2) as i16) as i16) as i32) as i16;
+    let x2 = (32767 - x2 as i32 + frac_mul16(x2, -7651 + frac_mul16(x2, 8277 + frac_mul16(-626, x2) as i16) as i16)) as i16;
     1 + x2
 }
